refactored eventForm to include the add availability and create event button from create-event
