+++ src/app/create-event/page.tsx
@@ -1,15 +1,13 @@
 'use client'
-import { useState, useRef, useEffect } from 'react'
-import { randomUUID, UUID } from 'crypto'
+import { useState, useEffect } from 'react'
+import { UUID } from 'crypto'
 import { addUserCreateEvent, getUser } from '@/utils/userUtils'
-import { insertEvent } from '@/utils/eventsUtils'
 import { addAttendee, Schedule } from '@/utils/attendeesUtils'
 import { useRouter } from 'next/navigation'
 
 import EventForm from '@/components/EventForm'
 import Grid from '@/components/AvailabilityGrid'
 import Header from '@/components/Header'
-import Username from '@/components/Username'
 
 export default function CreateEvent() {
   const [title, setTitle] = useState<string | null>('')
@@ -24,20 +22,19 @@ export default function CreateEvent() {
 
   const [isAvailable, setIsAvailable] = useState(false) // set to true when name is entered at sign in, Determines if the grid is selectable (selection mode)
   const [userName, setUserName] = useState<string | null>(null) // set to name entered at sign in
-  const dialogRef = useRef<HTMLDialogElement>(null) // modal
 
-  const [isButtonsVisible, setIsButtonsVisible] = useState(false) // New state to control visibility of buttons
   const [loading, setLoading] = useState(true)
 
   useEffect(() => {
     // Check if user is signed in
     const promises = []
     if (localStorage.getItem('username')) {
-      setUserName(localStorage.getItem('username') as string)
       const promise = getUser(localStorage.getItem('username') as UUID)
         .then((data) => {
           if (data) {
             setUserName(data[0].name)
+          } else {
+            setUserName('Guest')
           }
         })
         .catch((error) => {
@@ -140,17 +137,6 @@ export default function CreateEvent() {
     }
     // handleSaveResponse() // Call handleSaveResponse to save the response
     setIsAvailable(false) // Set availability to false when user creates event/saves their availability
-    setIsButtonsVisible(false) // Hide buttons after user creates event
-  }
-
-  // Function to open modal for after clicking "Sign In"
-  const openModal = () => {
-    if (dialogRef.current && !userName) {
-      dialogRef.current.showModal()
-    } else {
-      setIsAvailable(true)
-      setIsButtonsVisible(true)
-    }
   }
 
   return (
@@ -162,12 +148,9 @@ export default function CreateEvent() {
         <section //Left side container (Event form)
           className="sticky top-0 h-full w-full rounded-lg px-6 pb-16 shadow-lg md:w-[30%]"
         >
-          <div className="mb-6">
-            {userName != null && isAvailable && (
-              <Username username={userName} setUsername={setUserName} />
-            )}
-          </div>
           <EventForm
+            username={userName}
+            setUsername={setUserName}
             title={title}
             setTitle={setTitle}
             description={description}
@@ -184,50 +167,10 @@ export default function CreateEvent() {
             setConfig={setConfig}
             timezone={timezone}
             setTimezone={setTimezone}
+            isAvailable={isAvailable}
+            setIsAvailable={setIsAvailable}
+            handleSubmit={handleSubmit}
           />
-
-          <div //button container for positioning button
-            className="mx-4 flex justify-center pt-8"
-          >
-            {!isAvailable && ( //"Add Availability" button is only visible when user has not signed in and added their availability
-              <button
-                className="btn btn-primary ml-4 rounded-full px-4 py-2"
-                onClick={openModal}
-              >
-                Add Availability
-              </button>
-            )}
-
-            <dialog ref={dialogRef} id="username_modal" className="modal">
-              <div className="modal-box focus:outline-white ">
-                <h3 className="py-4 text-lg font-bold">Sign In</h3>
-
-                <input
-                  type="text"
-                  placeholder="Enter Your Name"
-                  className="input input-bordered w-full max-w-xs py-4"
-                  value={userName ? userName : ''}
-                  onChange={(e) => {
-                    setUserName(e.target.value)
-                  }}
-                />
-
-                <div className="modal-action">
-                  <form method="dialog">
-                    <button
-                      className="btn btn-primary ml-4 rounded-full px-4 py-2"
-                      onClick={() => {
-                        setIsAvailable(true) // Update isAvailable to true when name is entered
-                        setIsButtonsVisible(true) // Show buttons when user signs in
-                      }}
-                    >
-                      Sign In
-                    </button>
-                  </form>
-                </div>
-              </div>
-            </dialog>
-          </div>
         </section>
 
         <section //Middle side container (Availability Grid)
@@ -242,19 +185,6 @@ export default function CreateEvent() {
             schedule={schedule}
             setSchedule={setSchedule}
           />
-
-          {isButtonsVisible && ( // Conditionally render buttons section
-            <div //button container for positioning "Create Event" button
-              className="flex flex-row justify-center gap-4 pt-8 "
-            >
-              <button
-                className="btn btn-primary rounded-full px-4 py-2"
-                onClick={handleSubmit}
-              >
-                Create Event
-              </button>
-            </div>
-          )}
         </section>
       </div>
     </div>
